+++ src/full/Agda/TypeChecking/Monad/Base.hs
@@ -5409,9 +5409,6 @@ data TypeError
     -- Positivity and polarity errors
         | DatatypeIndexPolarity
             -- ^ An index of a data type has a polarity different from 'Mixed'.
-        | RecursiveRecordNeedsInductivity QName
-            -- ^ A record type inferred as recursive needs a manual declaration
-            --   whether it should be inductively or coinductively.
 
     -- Sized type errors
         | CannotSolveSizeConstraints (List1 (ProblemConstraint, HypSizeConstraint)) Doc